Extract A11yContainerTest
